feat: disperse contract function working on hardhat local node and added ui small enhancements
