+++ packages/nextjs/components/Step3.tsx
@@ -1,7 +1,13 @@
+"use client";
+
 import React, { SyntheticEvent, useState } from "react";
 import Next from "./Next";
 import { Title } from "./Title";
 import { Variants, motion } from "framer-motion";
+import { isAddress } from "viem";
+import { createPublicClient, http } from "viem";
+import { mainnet } from "viem/chains";
+import { normalize } from "viem/ens";
 import { useGlobalState } from "~~/services/store/store";
 
 const pageVariant: Variants = {
@@ -40,9 +46,16 @@ const pageVariant: Variants = {
   },
 };
 
+const publicClient = createPublicClient({
+  chain: mainnet,
+  transport: http(),
+});
+
 export default function Step3({ handleNext, handleBack }: { handleNext: any; handleBack: any }) {
   const [goback, setGoBack] = useState<boolean>(false);
   const { disperseFormData, setDisperseFormData } = useGlobalState();
+
+  const [showAttesterENS, setShowAttesterENS] = useState(false);
 
   const catchSubmit = (e: SyntheticEvent) => {
     e.preventDefault();
@@ -59,6 +72,24 @@ export default function Step3({ handleNext, handleBack }: { handleNext: any; han
       }
     }
   };
+
+  const handleInputChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (isAddress(e.target.value)) {
+      setDisperseFormData("attesterAddress", e.target.value);
+    } else {
+      setDisperseFormData("attesterENS", e.target.value);
+      setShowAttesterENS(true);
+
+      const ensAddress = await publicClient.getEnsAddress({
+        name: normalize(e.target.value),
+      });
+
+      if (ensAddress) {
+        setDisperseFormData("attesterAddress", ensAddress);
+      }
+    }
+  };
+
   return (
     <form onSubmit={catchSubmit} className=" w-full flex flex-col  items-center md:h-[580px] md:justify-between">
       <motion.section
@@ -68,15 +99,15 @@ export default function Step3({ handleNext, handleBack }: { handleNext: any; han
         exit={goback ? "exit2" : "exit"}
         className="flex flex-col  mb-8 md:mb-0 bg-white w-[90%] rounded-2xl py-10 px-7 z-30 relative bottom-24 text-[14px] md:bottom-0 md:p-0 md:w-[70%] h-full"
       >
-        <Title title="Enter attester address">
-          Enter the attester address you want to validate the attestations from.
+        <Title title="Enter attester address or ENS">
+          Enter the attester address or ENS you want to validate the attestations from.
         </Title>
         <div className="w-full flex flex-col space-y-4 md:space-y-5">
           <input
             type="text"
             name="attesterAddress"
-            value={disperseFormData.attesterAddress}
-            onChange={event => setDisperseFormData("attesterAddress", event.target.value)}
+            value={!showAttesterENS ? disperseFormData.attesterAddress : disperseFormData.attesterENS}
+            onChange={handleInputChange}
             placeholder="Enter attester address"
             className="mt-4 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
           />
